requirements.txt
asgiref==3.5.2
Django==4.0.6
django-restframework==0.0.1
djangorestframework==3.13.1
numpy==1.23.0
opencv-python==4.6.0.66
pytz==2022.1
scipy==1.8.1
sqlparse==0.4.2
